requirements.txt
gspread==5.12.0
oauth2client==4.1.3
pillow==10.1.0
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0
facebook-sdk==3.1.